refactor: extract focus.c from somewm.c
Move focus policy into a dedicated module: focusclient (the main
focus function handling raise, unfocus old, activate new, keyboard
enter, signal emission), focustop (find topmost client on monitor),
and some_update_pointer_constraint.

Replace scattered extern declarations for focusclient/focustop in
input.c, monitor.c, protocols.c, window.c, and somewm_api.c with
#include "focus.h".

somewm.c: 1811 -> 1608 lines (-203)
focus.c: 240 lines (new)
