+++ meson.build
@@ -289,6 +289,7 @@ somewm_sources = [
   'monitor.c',
   'input.c',
   'window.c',
+  'focus.c',
   'ipc.c',
   'color.c',
   'draw.c',
